requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.5.0=py310h5588dad_0
asgiref=3.5.0=pyhd8ed1ab_0
attrs=21.4.0=pypi_0
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.10.8=h5b45459_0
cached-property=1.5.2=pypi_0
certifi=2021.10.8=pypi_0
cffi=1.15.0=py310hcbf9ad4_0
charset-normalizer=2.0.12=pypi_0
click=8.1.2=py310h5588dad_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.2=py310ha857299_0
ecdsa=0.17.0=pyhd8ed1ab_0
fastapi=0.75.0=pyhd8ed1ab_0
h11=0.13.0=pyhd8ed1ab_0
idna=3.3=pyhd8ed1ab_0
isodate=0.6.1=pypi_0
libffi=3.4.2=h8ffe710_5
libzlib=1.2.11=h8ffe710_1014
lxml=4.8.0=pypi_0
openssl=1.1.1n=h8ffe710_0
pip=22.0.4=pyhd8ed1ab_0
platformdirs=2.5.1=pypi_0
pyasn1=0.4.8=py_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.9.0=py310he2412df_0
python=3.10.4=h9a09f29_0_cpython
python-jose=3.3.0=pyh6c4a22f_1
python-multipart=0.0.5=py_0
python_abi=3.10=2_cp310
pytz=2022.1=pypi_0
requests=2.27.1=pypi_0
requests-file=1.5.1=pypi_0
requests-toolbelt=0.9.1=pypi_0
rsa=4.8=pyhd8ed1ab_0
setuptools=61.3.0=py310h5588dad_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=py310h5588dad_3
sqlite=3.37.1=h8ffe710_0
starlette=0.17.1=pyhd8ed1ab_0
studentvue=2.0.5=pypi_0
tk=8.6.12=h8ffe710_0
typing-extensions=4.1.1=hd8ed1ab_0
typing_extensions=4.1.1=pyha770c72_0
tzdata=2022a=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.9=pypi_0
uvicorn=0.17.6=py310h5588dad_0
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
wheel=0.37.1=pyhd8ed1ab_0
xmljson=0.2.1=pypi_0
xz=5.2.5=h62dcd97_1
zeep=4.1.0=pypi_0
